Drop unused code and move translate tests
