+++ controller/hybridgateway/builder/kongplugin.go
@@ -6,9 +6,7 @@ import (
 	"fmt"
 	"maps"
 
-	"github.com/samber/lo"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	configurationv1 "github.com/kong/kong-operator/api/configuration/v1"
 	"github.com/kong/kong-operator/controller/hybridgateway/metadata"
@@ -112,269 +110,4 @@ func (b *KongPluginBuilder) WithPluginName(name string) *KongPluginBuilder {
 func (b *KongPluginBuilder) WithPluginConfig(config json.RawMessage) *KongPluginBuilder {
 	b.plugin.Config.Raw = config
 	return b
-}
-
-// WithFilter sets the KongPlugin configuration based on the given HTTPRouteFilter.
-func (b *KongPluginBuilder) WithFilter(filter gwtypes.HTTPRouteFilter) *KongPluginBuilder {
-
-	switch filter.Type {
-	case gatewayv1.HTTPRouteFilterRequestHeaderModifier:
-		rt, err := translateRequestModifier(filter)
-		if err != nil {
-			b.errors = append(b.errors, err)
-			return b
-		}
-
-		b.plugin.PluginName = "request-transformer"
-
-		configJSON, err := json.Marshal(rt)
-		if err != nil {
-			b.errors = append(b.errors, fmt.Errorf("failed to marshal %q plugin config: %w", b.plugin.PluginName, err))
-			return b
-		}
-		b.plugin.Config.Raw = configJSON
-	case gatewayv1.HTTPRouteFilterResponseHeaderModifier:
-		rt, err := translateResponseModifier(filter)
-		if err != nil {
-			b.errors = append(b.errors, err)
-			return b
-		}
-
-		b.plugin.PluginName = "response-transformer"
-
-		configJSON, err := json.Marshal(rt)
-		if err != nil {
-			b.errors = append(b.errors, fmt.Errorf("failed to marshal %q plugin config: %w", b.plugin.PluginName, err))
-			return b
-		}
-		b.plugin.Config.Raw = configJSON
-	case gatewayv1.HTTPRouteFilterRequestRedirect:
-		pluginConf, err := translateRequestRedirect(filter)
-		if err != nil {
-			b.errors = append(b.errors, err)
-			return b
-		}
-
-		b.plugin.PluginName = "redirect"
-
-		configJSON, err := json.Marshal(pluginConf)
-		if err != nil {
-			b.errors = append(b.errors, fmt.Errorf("failed to marshal %q plugin config: %w", b.plugin.PluginName, err))
-			return b
-		}
-		b.plugin.Config.Raw = configJSON
-	case gatewayv1.HTTPRouteFilterURLRewrite:
-		pluginConf, err := translateURLRewrite(filter)
-		if err != nil {
-			b.errors = append(b.errors, err)
-			return b
-		}
-
-		b.plugin.PluginName = "request-transformer"
-
-		configJSON, err := json.Marshal(pluginConf)
-		if err != nil {
-			b.errors = append(b.errors, fmt.Errorf("failed to marshal %q plugin config: %w", b.plugin.PluginName, err))
-			return b
-		}
-		b.plugin.Config.Raw = configJSON
-	default:
-		b.errors = append(b.errors, fmt.Errorf("unsupported filter type: %s", filter.Type))
-	}
-	return b
-}
-
-// internal functions and types for translating HTTPRouteFilter to KongPlugin configurations
-
-type transformerTargetSlice struct {
-	Headers []string `json:"headers,omitempty"`
-}
-
-type transformerTargetSliceReplace struct {
-	transformerTargetSlice
-	Uri string `json:"uri,omitempty"`
-}
-
-type transformerData struct {
-	// Add: adds an header only if the header is not existent.
-	// Append: adds a new header even if the header is already existent (adds a new instance).
-	// Remove: removes an entry.
-	// Replace: overwrites an header value only if the header exists.
-	Add     transformerTargetSlice        `json:"add,omitzero"`
-	Append  transformerTargetSlice        `json:"append,omitzero"`
-	Remove  transformerTargetSlice        `json:"remove,omitzero"`
-	Replace transformerTargetSliceReplace `json:"replace,omitzero"`
-}
-
-func translateRequestModifier(filter gwtypes.HTTPRouteFilter) (transformerData, error) {
-	var err error
-	plugin := transformerData{}
-
-	if filter.RequestHeaderModifier == nil {
-		err = errors.New("RequestHeaderModifier filter config is missing")
-		return plugin, err
-	}
-
-	// In order to overwrite an header of add if not present (GWAPI Set) we should do a Kong Plugin
-	// Replace (so it will overwrite it if found) + Add (so if not found, will add it).
-	if len(filter.RequestHeaderModifier.Set) > 0 {
-		for _, v := range filter.RequestHeaderModifier.Set {
-			plugin.Replace.Headers = append(plugin.Replace.Headers, string(v.Name)+":"+v.Value)
-			plugin.Add.Headers = append(plugin.Add.Headers, string(v.Name)+":"+v.Value)
-		}
-	}
-	// Add for GWAPI equals "append" for Kong Plugins (it will add another instance of the header).
-	if len(filter.RequestHeaderModifier.Add) > 0 {
-		for _, v := range filter.RequestHeaderModifier.Add {
-			plugin.Append.Headers = append(plugin.Append.Headers, string(v.Name)+" :"+v.Value)
-		}
-	}
-	if len(filter.RequestHeaderModifier.Remove) > 0 {
-		plugin.Remove.Headers = append(plugin.Remove.Headers, filter.RequestHeaderModifier.Remove...)
-	}
-
-	if len(plugin.Add.Headers)+len(plugin.Append.Headers)+
-		len(plugin.Remove.Headers)+len(plugin.Replace.Headers) == 0 {
-		err = errors.New("RequestHeaderModifier filter config is empty")
-		plugin = transformerData{}
-	}
-	return plugin, err
-}
-
-func translateResponseModifier(filter gwtypes.HTTPRouteFilter) (transformerData, error) {
-	var err error
-	plugin := transformerData{}
-
-	if filter.ResponseHeaderModifier == nil {
-		err = errors.New("ResponseHeaderModifier filter config is missing")
-		return plugin, err
-	}
-
-	if len(filter.ResponseHeaderModifier.Set) > 0 {
-		for _, v := range filter.ResponseHeaderModifier.Set {
-			plugin.Replace.Headers = append(plugin.Replace.Headers, string(v.Name)+":"+v.Value)
-			plugin.Add.Headers = append(plugin.Add.Headers, string(v.Name)+":"+v.Value)
-		}
-	}
-	if len(filter.ResponseHeaderModifier.Add) > 0 {
-		for _, v := range filter.ResponseHeaderModifier.Add {
-			plugin.Append.Headers = append(plugin.Append.Headers, string(v.Name)+":"+v.Value)
-		}
-	}
-	if len(filter.ResponseHeaderModifier.Remove) > 0 {
-		plugin.Remove.Headers = append(plugin.Remove.Headers, filter.ResponseHeaderModifier.Remove...)
-	}
-
-	if len(plugin.Add.Headers)+len(plugin.Append.Headers)+len(plugin.Remove.Headers)+len(plugin.Replace.Headers) == 0 {
-		err = errors.New("ResponseHeaderModifier filter config is empty")
-		plugin = transformerData{}
-	}
-	return plugin, err
-}
-
-type requestRedirectConfig struct {
-	KeepIncomingPath bool   `json:"keep_incoming_path"`
-	Location         string `json:"location"`
-	StatusCode       int    `json:"status_code"`
-}
-
-func translateRequestRedirect(filter gwtypes.HTTPRouteFilter) (requestRedirectConfig, error) {
-	rr := filter.RequestRedirect
-
-	if rr == nil {
-		return requestRedirectConfig{}, errors.New("RequestRedirect filter config is missing")
-	}
-
-	// GWAPI default status code is 302.
-	plugin := requestRedirectConfig{StatusCode: 302}
-
-	if rr.StatusCode != nil {
-		plugin.StatusCode = *rr.StatusCode
-	}
-
-	locHost := translateRequestRedirectHostname(rr)
-	locPath, err := translateRequestRedirectPath(rr)
-	if err != nil {
-		return requestRedirectConfig{}, err
-	}
-	if locPath == "" {
-		plugin.KeepIncomingPath = true
-		locPath = "/"
-	}
-
-	plugin.Location = locHost + locPath
-	return plugin, nil
-}
-
-func translateRequestRedirectHostname(rr *gatewayv1.HTTPRequestRedirectFilter) string {
-	if rr.Hostname == nil || *rr.Hostname == "" {
-		return ""
-	}
-
-	// when no scheme specified we assume `http` (as KIC does) but we should preserve the actual scheme instead.
-	// this cannot be done with direct filter -> kong plugin conversion.
-	// See https://github.com/Kong/kong-operator/issues/2466.
-	host := lo.FromPtrOr(rr.Scheme, "http") + "://"
-	host += string(lo.FromPtrOr((rr.Hostname), ""))
-	if rr.Port != nil {
-		host += fmt.Sprintf(":%d", *rr.Port)
-	}
-	return host
-}
-
-func translateRequestRedirectPath(rr *gatewayv1.HTTPRequestRedirectFilter) (string, error) {
-	path := ""
-	var err error
-
-	if rr.Path == nil {
-		return path, nil
-	}
-
-	pathModifier := rr.Path
-	switch pathModifier.Type {
-	case gatewayv1.FullPathHTTPPathModifier:
-		path = translateRequestRedirectPathFullPath(pathModifier.ReplaceFullPath)
-	case gatewayv1.PrefixMatchHTTPPathModifier:
-		path = translateRequestRedirectPathPrefixMatch(pathModifier.ReplacePrefixMatch)
-	default:
-		err = errors.New("unsupported RequestRedirect path modifier type: " + string(pathModifier.Type))
-	}
-	return path, err
-}
-
-func translateRequestRedirectPathFullPath(replaceFullPath *string) string {
-	if replaceFullPath == nil || *replaceFullPath == "" {
-		return "/"
-	}
-	return *replaceFullPath
-}
-
-func translateRequestRedirectPathPrefixMatch(prefixMatch *string) string {
-	// Not implemented yet - Kong does not have a direct equivalent for prefix match replacement.
-	// KIC in Konnect just ignores PrefixMatch filters, let's do the same.
-	// Tracker: https://github.com/Kong/kong-operator/issues/2466
-	return "/"
-}
-
-func translateURLRewrite(filter gwtypes.HTTPRouteFilter) (transformerData, error) {
-	ur := filter.URLRewrite
-	pluginConf := transformerData{}
-
-	if ur == nil {
-		return pluginConf, errors.New("URLRewrite filter config is missing")
-
-	}
-
-	if ur.Path != nil && ur.Path.Type == gatewayv1.FullPathHTTPPathModifier {
-		pluginConf.Replace.Uri = translateURLRewritePathFullPath(ur.Path.ReplaceFullPath)
-	}
-
-	return pluginConf, nil
-}
-
-func translateURLRewritePathFullPath(replaceFullPath *string) string {
-	if replaceFullPath == nil || *replaceFullPath == "" {
-		return "/"
-	}
-	return *replaceFullPath
 }
